% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{colinadummy}
\alias{colinadummy}
\title{Objeto \code{curvacolina} Exemplo}
\format{
Lista de um elemento chamado \code{CC}, um data.table de 1888 com as colunas
\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia gerada}
\item{\code{rend}}{rendimento correspondente}
}

Adicionalmente, contem os atributos

\describe{
\item{\code{ncurvas}}{numero de curvas na colina}
\item{\code{rends}}{vetor de rendimentos contidos na curva colina}
}
}
\usage{
colinadummy
}
\description{
Exemplo de objeto retornado por uma das funcoes de leitura \code{\link{learqcolina}} ou
\code{\link{learqprocit}}
}
\keyword{datasets}
