% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacao.r, R/mod-interpoladorM.r,
%   R/mod-tensorprod.r, R/mod-thinplate.r, R/mod-triangulacao.r,
%   R/retrocompat.r, R/utils.r
\name{getcolina.interpolador}
\alias{getcolina.interpolador}
\alias{getcolina.interpoladorM}
\alias{getcolina.tensorprod}
\alias{getcolina.thinplate}
\alias{getcolina.triangulacao}
\alias{getcolina.retamindist}
\alias{getcolina}
\title{Generica Para Extrair Colina Original do Ajuste}
\usage{
\method{getcolina}{interpolador}(object)

\method{getcolina}{interpoladorM}(object)

\method{getcolina}{tensorprod}(object)

\method{getcolina}{thinplate}(object)

\method{getcolina}{triangulacao}(object)

\method{getcolina}{retamindist}(object)

getcolina(object)
}
\arguments{
\item{object}{objeto do qual extrair \code{data.table} da curva colina. Pode ser um modelo
interpolador ou objeto da classe \code{curvacolina}}
}
\value{
\code{data.table} contendo a curva colina em formato padronizado
}
\description{
Extrai o dado ajustado corretamente dependendo do tipo de modelo. Funcao interna
}
