% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-tensorprod.r
\name{getcolina.tensorprod}
\alias{getcolina.tensorprod}
\title{Extracao Da Colina Original Em Objetos \code{triangulacao}}
\usage{
\method{getcolina}{tensorprod}(object)
}
\description{
Extracao Da Colina Original Em Objetos \code{triangulacao}
}
