% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-trangulacao.r
\name{getcolina.triangulacao}
\alias{getcolina.triangulacao}
\title{Extracao Da Colina Original Em Objetos \code{triangulacao}}
\usage{
\method{getcolina}{triangulacao}(object)
}
\description{
Extracao Da Colina Original Em Objetos \code{triangulacao}
}
