% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradecolina.r
\name{gradecolina}
\alias{gradecolina}
\title{Objetos \code{gradecolina}}
\description{
Tabelas regulares amostradas de um objeto \code{\link{interpolador}}.
}
\details{
A chamada do metodo \code{predict} em objetos da classe \code{interpolador} e suas subclasses
pode ser retornada de forma mais completa. Esta forma consiste na tabela regularmente espacada
para representacao discreta da curva colina.

Objetos do tipo \code{gradecolina} sao uma lista de dois elementos, o primeiro dos quais um
data.table com as colunas

\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia gerada}
\item{\code{rend}}{rendimento interpolado}
\item{\code{inhull}}{booleano indicando se o ponto foi interpolado (\code{TRUE}) ou extrapolado (\code{FALSE})}
}

O segundo elemento e um objeto \code{curvacolina}, contendo a colina original.
}
\seealso{
Other gradecolina: 
\code{\link{interpolacao_bilinear}},
\code{\link{plot.gradecolina}()}
}
\concept{gradecolina}
