% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{INTERPBILIN}
\alias{INTERPBILIN}
\title{Interpolacao Bilinear}
\usage{
INTERPBILIN(hlGrade, potGrade, rendGrade, hlPred, potPred)
}
\arguments{
\item{hlGrade, potGrade, rendGrade}{valores das variaveis explicativas (x e y) e dependente (z), na grade}

\item{hlPred, potPred}{valores das variaveis explicativas a interpolar na grade}
}
\value{
vetor de valores interpolados
}
\description{
Realiza interpolacao bilinear em uma grade tridimensional de valores. Funcao interna
}
\details{
Os argumentos \code{hlGrade} e \code{potGrade} devem corresponder as coordenadas de queda e
potencia contidas na grade em ordem crescente. \code{rendGrade} deve ser uma matrix M x N
com o formato\tabular{lllll}{
    \tab pot_1 \tab pot_2 \tab ... \tab pot_N \cr
   hl_1 \tab rend_11 \tab rend_12 \tab ... \tab rend_1N \cr
   hl_2 \tab rend_21 \tab rend_22 \tab ... \tab rend_2N \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
   hl_M \tab rend_M1 \tab rend_M2 \tab ... \tab rend_MN \cr
}


\code{hlPred} e \code{potPred} devem ser informados como vetores numericos indicando as
coordenadas onde interpolar. Os pontos devem ser fornecidos ordenados por \code{potPred} em
ordem crescente. Isto e necessario para que se possa usar uma pesquisa mais otimizada na matriz
de rendimentos e, assim, tornar a funcao mais eficiente
}
