% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacao.r
\name{interpolador}
\alias{interpolador}
\title{Modelo Para Interpolacao De Curva Colina}
\usage{
interpolador(
  colina,
  metodo = c("triangulacao", "thinplate", "tensorprod"),
  ...
)
}
\arguments{
\item{colina}{objeto da classe \code{curvacolina} (retornado pelas funcoes de leitura)}

\item{metodo}{um de \code{c("triangulacao", "thinplate", "tensorpro")}. Ver Detalhes}

\item{...}{demais parametros que possam ser passados as funcoes de ajuste de cada \code{metodo}.
Ver Detalhes}
}
\value{
objeto da classe \code{interpolador} e subclasse \code{metodo}, isto e, um modelo com o
qual se realizar previsoes e, assim, interpolar o dado original
}
\description{
Funcao para estimacao de diferentes interpoladores de curva colina
}
\details{
\code{interpolador} serve como uma interface comum para estimacao de diversas formas atraves das
quais interpolar a colina original. Atualmente ha tres abordagens implementadas:

\itemize{
\item{\code{"triangulacao"}}
\item{\code{"thinplate"}}
\item{\code{"tensorprod"}}
}

Se \code{metodo = "triangulacao"}, a colina fornecida e projetada no plano hl x pot e entao
tesselada atraves da triangulacao de Delaunay. Interpolacao de um ponto em objetos de
triangulacao se da atraves de transformacao para coordenadas baricentricas e subsequente media
ponderada dos rendimentos nos vertices do triangulo que o contem.

Quando \code{metodo = "thinplate"} ou \code{metodo = "tensorprod"}, a colina e suavizada atraves
de splines, cujo tipo depende do \code{metodo} passado: \code{"thinplate"} corresponde a
suavizacao por splines homonimas e \code{"tensorprod"} ao produto tensor de splines. A
interpolacao nesse caso consiste simplesmente da consulta a curva suavizada em pontos
selecionados.

O argumento \code{...} so tem uso quando \code{metodo} corresponde a uma suavizacao por splines.
Nestes casos, \code{...} pode conter qualquer um dos argumentos passados ao construtor do modelo,
isto, e, tipo de spline (\code{bs}) e dimensao da base (\code{k}). Ver (\code{\link[mgcv]{te}}).
}
\examples{

# usando dado dummy contido no pacote
interp_tri <- interpolador(colinadummy, "tri")

# interpolando uma grade 20x20 no dominio da colina
pontos <- geragrade(colinadummy, 20, 20)
interp <- predict(interp_tri, pontos)

}
\seealso{
Other interpolador: 
\code{\link{plot.interpolador}()},
\code{\link{predict.interpolador}()}
}
\concept{interpolador}
