% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradecolina.r
\name{interpolagrade}
\alias{interpolagrade}
\alias{interpolagrade.data.frame}
\alias{interpolagrade.curvacolina}
\title{Interpolacao Bilinear}
\usage{
interpolagrade(pontos, gradecolina, full.output)

\method{interpolagrade}{data.frame}(pontos, gradecolina, full.output = FALSE)

\method{interpolagrade}{curvacolina}(pontos, gradecolina, full.output = FALSE)
}
\arguments{
\item{pontos}{\code{curvacolina} ou data.frame-like contendo as coordenadas \code{hl} \code{pot}
dos pontos a interpolar. Ver Detalhes}

\item{gradecolina}{data.frame-like contendo colunas \code{hl}, \code{pot} e \code{rend}}

\item{full.output}{booleano -- se \code{FALSE} (padrao) retorna apenas o vetor de rendimentos
interpolados nas coordenadas \code{pontos}; se \code{TRUE} um data.table de \code{pontos} com
a coluna \code{rend} adicionada}
}
\value{
set \code{full.output = FALSE} vetor de rendimentos interpolados, do contrario um
\code{data.table} contendo \code{pontos} adicionado da coluna \code{rend} com resultado da
interpolacao
}
\description{
Interpola \code{pontos} na grade bivariada \code{gradecolina}
}
\details{
\code{pontos} permite que seja passado um objeto \code{curvacolina} para facilitar a interpolacao
dos pontos originais numa tabela gerada por algum dos metodos de interpolacao. Do contratio,
\code{pontos} deve ser um data.frame ou data.table com as colunas \code{hl} e \code{pot} contendo
as coordenadas para interpolar.
}
\examples{

# usando o interpolador de triangulacao
tri <- interpolador(colinadummy, "tri")

# extrai uma grade dele
coord <- geragrade(colinadummy, 10, 10)
gradecolina <- predict(tri, coord, full.output = TRUE)

\dontrun{
# interpolacao de pontos da colina na grade
interpolagrade(colinadummy, gradecolina)

# interpolacao retornando dado cheio
interpolagrade(colinadummy, gradecolina, full.output = TRUE)
}

}
