% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{new_curvacolina}
\alias{new_curvacolina}
\title{Construtor Interno Da Classe \code{curvacolina}}
\usage{
new_curvacolina(curvas, g, rho)
}
\arguments{
\item{curvas}{data.table contendo a digitalizacao da colina. Ver Detalhes}

\item{g}{aceleracao da gravidade}

\item{rho}{massa especifica da agua}
}
\value{
objeto \code{curvacolina}, uma lista de um elemento \code{data.table} contendo as colunas

\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia gerada}
\item{\code{vaz}}{vazao turbinada}
\item{\code{rend}}{rendimento correspondente}
}

Adicionalmente tem os atributos:

\describe{
\item{\code{ncurvas}}{numero de curvas na colina}
\item{\code{rends}}{vetor de rendimentos contidos na curva colina}
\item{\code{max}}{rendimento no "olho" da colina}
\item{\code{rho}}{densidade da agua associada}
\item{\code{g}}{gravidade associada}
}
}
\description{
Construtor para uso interno, nao deve ser chamado diretamente pelo usuario
}
\details{
O dado \code{curvas} deve ser um data.table de tres colunas nomeadas

\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia na turbina}
\item{\code{rend}}{rendimento da turbina}
}

Conjuntamente com os valores de \code{rho} e \code{g} este dado e utilizado para calculo da
vazao. Se nao forem passados, serao assumidos como NA e a vazao tambem sera NA.

Caso ja exista uma coluna de vazao chamada \code{vaz}, ela sera preservada no dado. Caso sejam
passados \code{rho} e \code{g} nesse caso eles ficarao incompativeis com os valores de vazao.
}
