% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gradecolina.r
\name{new_gradecolina}
\alias{new_gradecolina}
\title{Construtor Interno De \code{gradecolina}}
\usage{
new_gradecolina(pontos, rends, interpolador)
}
\arguments{
\item{pontos}{data.frame-like contendo coordenadas interpoladas}

\item{rends}{vetor numerico de rendimentos interpolados}

\item{interpolador}{tipo de interpolador de onde foi extraida}
}
\value{
objeto da classe \code{gradecolina}; lista de dois elementos, o primeiro dos quais um
data.table com as colunas

\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia gerada}
\item{\code{rend}}{rendimento interpolado}
\item{\code{inhull}}{booleano indicando se o ponto foi interpolado (\code{TRUE}) ou extrapolado (\code{FALSE})}
}

O segundo elemento e um objeto \code{curvacolina}, contendo a colina original
}
\description{
Funcao interna do pacote, nao deve ser chamada pelo usuario
}
