% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-interpoladorM.r
\name{new_interpoladorM}
\alias{new_interpoladorM}
\title{Construtor Interno De \code{interpoladorM}}
\usage{
new_interpoladorM(modelos, quebra, modo)
}
\arguments{
\item{modelos}{lista de interpoladores estimados}

\item{quebra}{numerico indicando o rendimento da curva a partir da qual troca de interpolador}
}
\value{
objeto da classe \code{interpolador} com subclasse \code{"interpoladorM"}
}
\description{
Funcao interna que nao deve ser chamada pelo usuario
}
