% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{orderpoly}
\alias{orderpoly}
\title{Ordem Dos Vertices De Poligono}
\usage{
orderpoly(dat, centro_hl, centro_pot)
}
\arguments{
\item{dat}{\code{data.table} contendo vertices de um poligono cuja ordem sera determinada}

\item{centro_hl}{opcional, valor para centralizar \code{dat$hl}. Caso vazio usa a media}

\item{centro_pot}{opcional, valor para centralizar \code{dat$pot}. Caso vazio usa a media}
}
\value{
vetor de inteiros indicando os indices dos vertices na ordenacao por angulo
}
\description{
Centraliza \code{dat} por referencia e retorna vetor de indices dos vertices em ordem adjacente
}
\details{
A determinacao de ordem dos vertices e feita com base no angulo entre cada um e o centro do
poligono. Este centro pode ser infromado por meio dos argumentos \code{centro_hl} e
\code{centro_pot} ou, caso estes fiquem vazios, sera calculado como a media amostral.
}
