% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{parsedadocolina}
\alias{parsedadocolina}
\title{Parser De Dados Para \code{as.curvacolina}}
\usage{
parsedadocolina(x, force)
}
\arguments{
\item{x}{\code{data.frame}-like a ser convertido}

\item{force}{booleano indicando se o objeto deve ser forcado ao formato padrao. Pode gerar
resultados inconsistentes. Ver Detalhes}
}
\value{
dado parsed para transformacao
}
\description{
Funcao interna, nao deve ser chamada pelo usuario
}
\details{
Esta funcao executa todos os checks e coercoes necessarias para execucao da transformacao
\code{as.curvacolina}
}
