% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.curvacolina}
\alias{plot.curvacolina}
\title{Visualizacao De \code{curvacolina}}
\usage{
\method{plot}{curvacolina}(x, tipo = c("3d", "2d"), print = TRUE, modo = "pot", ...)
}
\arguments{
\item{x}{objeto \code{curvacolina} retornado por uma das funcoes de leitura}

\item{tipo}{um de \code{c("3d", "2d")} indicando o tipo de grafico desejado}

\item{print}{booleano indicando se o plot deve ser exibido. Caso \code{print = FALSE} o objeto
sera retornado silenciosamente}

\item{modo}{um de \code{c("pot", "vaz")}, indicando qual o modo de curva colina esta sendo}

\item{...}{existe somente para consistencia de metodos. Nao possui utilidade}
}
\value{
se \code{tipo = "3d"} um objeto \code{plotly}, do contrario um objeto \code{ggplot}
contendo o plot. Em ambos os casos o grafico so sera exibido ao usuario caso
\code{print = TRUE} (o padrao).
}
\description{
Plots bi ou tridimensionais de curvas colinas
}
\examples{

arq <- system.file("extdata/procit_cc_alterada.xlsx", package = "curvacolina")
colina <- learqcolina(arq)

\dontrun{
# plot 3d
plot(colina[[1]], "3d")
plot(colina[[1]], "2d")

# execucao silenciosa e posterior exibicao
p <- plot(colina, "2d", print = FALSE)
print(p)

# plotando por vazao
plot(colina[[1]], "3d", modo = "vaz")
plot(colina[[1]], "2d", modo = "vaz")

}

}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{coordgrade}()},
\code{\link{learqcolina}()},
\code{\link{summary.curvacolina}()},
\code{\link{write.curvacolina}()}
}
\concept{curvacolina}
