% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.interpolador}
\alias{plot.interpolador}
\title{Visualizacao De Interpolacoes}
\usage{
\method{plot}{interpolador}(
  x,
  tipo = c("3d", "2d"),
  add_colina = TRUE,
  hl = 200,
  pot = 200,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{objeto \code{interpolador} retornado pela funcao homonima}

\item{tipo}{um de \code{c("3d", "2d")} indicando o tipo de grafico desejado}

\item{add_colina}{booleano indicando se os pontos da colina original tambem devem ser plotados}

\item{hl, pot}{vetores definindo grade amostrada. Ver Detalhes}

\item{print}{booleano indicando se o plot deve ser exibido. Caso \code{print = FALSE} o objeto
sera retornado silenciosamente}

\item{...}{existe somente para consistencia de metodos. Nao possui utilidade}
}
\value{
Se \code{tipo = "3d"} um objeto \code{plotly} contendo o plot 3d; se \code{tipo = "2d"}
um objeto \code{ggplot} contendo o plot 2d. Em ambos os casos o grafico so sera exibido ao
usuario caso \code{print = TRUE} (o padrao).
}
\description{
Funcao para visualizacao dos modelos ajustados atraves de \code{interpolador}
}
\details{
Os argumentos \code{hl} e \code{pot} devem ser fornecidos como um numero, indicando quantos
cortes no eixo de queda e potencia, respectivamente, devem ser utilizados na geracao de uma grade
a qual interpolar na superficie suavizada. Estes pontos serao entao utilizados para plot.
}
\seealso{
Other interpolador: 
\code{\link{interpolador}()},
\code{\link{predict.interpolador}()}
}
\concept{interpolador}
