% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PTINPOLY}
\alias{PTINPOLY}
\title{Implementacao Do Algoritmo Raycasting}
\usage{
PTINPOLY(points, poly)
}
\arguments{
\item{points}{matriz contendo as coordenadas dos pontos a testar}

\item{poly}{matriz contendo as coordenadas dos vertices do poligono. Ver Detalhes}
}
\value{
vetor coluna de 0s e 1s indicando quais pontos estao dentro do poligono
}
\description{
Identifica quais observacoes em \code{pontos} estao dentro do poligono \code{poly}
}
\details{
Esta funcao assume que linhas de \code{poly} representam vertices adjacentes. Nao importa onde
o primeiro vertice esta, contanto que cada nova linha indique o proximo vertice do poligono em
um mesmo sentido de deslocamento. Sera considerado que o primeiro vertice e o adjacente ao
ultimo no sentido de deslocamento.
}
