% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{reduzcolina}
\alias{reduzcolina}
\title{Redutor De \code{curvacolina}}
\usage{
reduzcolina(colina, taxa)
}
\arguments{
\item{colina}{objeto \code{curvacolina} retornado pelas funcoes de leitura}

\item{taxa}{inteiro indicando reducao do tamanho do historico em \code{taxa_reducao} vezes. Ver
Detalhes}
}
\value{
\code{colina} passada com observacoes selecionadas
}
\description{
Retorna o objeto \code{curvacolina} com um numero reduzido de observacoes
}
\details{
O argumento \code{taxa} e utilizado da seguinte maneira: seja \code{n} o numero de observacoes em
\code{colina}, sera amostrado um subset dessas tomando uma a cada \code{taxa} observacoes. Por
exemplo, se \code{taxa = 3}, sao selecionadas as observacoes 1, 4, 7, e assim por diante.

O processo de selecao descrito acima e realizado \bold{por curva de rendimento}, dado que a curva
tenha pelo menos \code{5 * taxa} observacoes. Este limite e algo essencialmente arbitrario.
}
