% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrocompat.r
\name{retamindist}
\alias{retamindist}
\title{Classe \code{retamindist}}
\usage{
retamindist(arq)
}
\arguments{
\item{arq}{caminho da grade gerada por interpolacao de reta com menor distancia}
}
\value{
objeto da classe \code{interpolador} e subclasse \code{retamindist} contendo a colina
interpolada pelo metodo antigo
}
\description{
Construtor de objetos da classe \code{retamindist}
}
\details{
A interpolacao antiga esta implementada em VBA e nao deve ser portada para outras linguagens. A
partir de execucoes especiais da macro e possivel gerar uma tabela de colina interpolada tal qual
as funcoes fornecidas por este pacote.

\code{retamindist} e uma funcao de leitura para importar as tabelas geradas e encapsular o dado
em uma classe que se comporte de forma similar aos outros interpoladores deste pacote
}
