% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{set_grho}
\alias{set_grho}
\title{Atribui \code{g} E \code{rho} A Objeto \code{curvacolina}}
\usage{
set_grho(x, g, rho)
}
\arguments{
\item{x}{objeto da classe \code{curvacolina}}

\item{g, rho}{aceleracao da gravidade e densidade da agua}
}
\value{
objeto \code{curvacolina} passado com atributos \code{g} e \code{rho}, assim como vazao
turbinada caclulada
}
\description{
Adiciona os atributos e calcula vazao turbinada na transformacao de curva colina
}
\examples{

\dontrun{
# o dado padrao do pacote nao possui g e rho
is.na(attr(colinadummy, "g"))
is.na(attr(colinadummy, "rho"))
}

# adicionando os valores de literatura
cc <- set_grho(colinadummy, 9.81, 1000)

\dontrun{
# atributos adicionados
attr(cc, "g")
attr(cc, "rho")

# vazao calculada
cc$CC$vaz
}

}
