% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{summary.curvacolina}
\alias{summary.curvacolina}
\title{Sumario De \code{curvacolina}}
\usage{
\method{summary}{curvacolina}(object, ...)
}
\arguments{
\item{object}{objeto \code{curvacolina}}

\item{...}{existe somente para consistencia de metodos. Nao possui utilidade}
}
\value{
imprime um sumario da curva colina, sem retornar nada
}
\description{
Produz um breve sumario da curva colina
}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{coordgrade}()},
\code{\link{learqcolina}()},
\code{\link{plot.curvacolina}()},
\code{\link{write.curvacolina}()}
}
\concept{curvacolina}
