% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-tensorprod.r
\name{tensorprod}
\alias{tensorprod}
\title{Suavizacao De \code{curvacolina} Por Produto Tensor De P-Splines}
\usage{
tensorprod(colina, modo = "pot", ...)
}
\arguments{
\item{colina}{objeto \code{curvacolina} retornado pelas funcoes de leitura}

\item{modo}{um de \code{c("pot", "vaz")}, indicando qual o modo de curva colina esta sendo
modelada}

\item{...}{parametros extras para configuracao do modelo estimado. Ver Detalhes em
\code{\link{interpolador}}}
}
\value{
objeto da classe \code{tensorprod} contendo a suavizacao da curva colina
}
\description{
Funcao interna executada quando \code{metodo = "tensorprod"} em \code{interpolador}
}
\details{
Esta funcao nao deve ser chamada pelo usuario diretamente na maioria dos casos
}
