% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-triangulacao.r
\name{tessdelaunay}
\alias{tessdelaunay}
\title{Triangulacao Delauney}
\usage{
tessdelaunay(colina, modo)
}
\arguments{
\item{colina}{objeto \code{curvacolina} contendo curva a tesselar}
}
\value{
matriz com tres colunas indicando, em cada linha, o indice em \code{colina$CC} dos pontos
correspondentes aos vertices de cada triangulo gerado
}
\description{
Realiza triangulacao do espaco pelo metodo de Delauney via \code{geometry::delaunayn}
}
