% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-triangulacao.r
\name{tessradial}
\alias{tessradial}
\title{Triangulacao Radial}
\usage{
tessradial(colina, modo)
}
\arguments{
\item{colina}{um data.frame ou data.table contendo queda liquida, potencia e rendimento}
}
\value{
matriz de tres colunas indicando o indicie em \code{dat} dos pontos correspondentes aos
vertices de cada triangulo. Cada linha corresponde a um triangulo
}
\description{
Define os triangulos sempre usando o maximo como um dos vertices
}
\details{
Este e um metodo de triangulacao especifico para uso em torno do maximo. Cada triangulo e
definido usando sempre o maximo como um dos vertices. Os dois vertices restantes sao pontos
adjacentes entre si na curva de rendimento imediatamente inferior ao maximo, ou seja, o espaco
entre ultima curva e ponto maximo e fatiado em triangulos com arestas radiais, para todos os
pontos da ultima curva.
}
