% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-thinplate.r
\name{thinplate}
\alias{thinplate}
\title{Suavizacao De \code{curvacolina} Thin Plate Regression Splines}
\usage{
thinplate(colina, taxa_reducao = 1)
}
\arguments{
\item{colina}{objeto \code{curvacolina} retornado pelas funcoes de leitura}

\item{taxa_reducao}{inteiro indicando reducao do tamanho do historico em \code{taxa_reducao}
vezes. Ver Detalhes}
}
\value{
objeto da classe \code{thinplate} contendo a suavizacao da curva colina
}
\description{
Funcao interna executada quando \code{metodo = "thinplate"} em \code{interpolador}
}
\details{
Esta funcao nao deve ser chamada pelo usuario diretamente na maioria dos casos
}
