% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{write.curvacolina}
\alias{write.curvacolina}
\title{Escrita De \code{curvacolina}}
\usage{
write.curvacolina(x, file)
}
\arguments{
\item{x}{objeto \code{curvacolina} a ser escrito}

\item{file}{caminho para escrita. Deve possuir extensao -- caso nao possua, sera adicionada}
}
\value{
Escreve \code{x} no caminho especificado como um csv de quatro colunas
}
\description{
Metodo para facilitacao de escrita de \code{curvacolina} lida pelas funcoes do pacote
}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{coordgrade}()},
\code{\link{learqcolina}()},
\code{\link{plot.curvacolina}()},
\code{\link{summary.curvacolina}()}
}
\concept{curvacolina}
