% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gradecolina.r
\name{write.gradecolina}
\alias{write.gradecolina}
\title{Escrita De \code{gradecolina}}
\usage{
write.gradecolina(x, file)
}
\arguments{
\item{x}{objeto \code{gradecolina} a ser escrito}

\item{file}{caminho para escrita com extensao de aquivo}
}
\value{
Escreve grade em \code{x} no caminho especificado
}
\description{
Metodo para facilitacao de escrita de \code{gradecolina} lida pelas funcoes do pacote
}
\seealso{
Other gradecolina: 
\code{\link{gradecolina}},
\code{\link{interpolacao_bilinear}},
\code{\link{plot.gradecolina}()}
}
\concept{gradecolina}
