% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{as.curvacolina}
\alias{as.curvacolina}
\title{Conversor Para \code{curvacolina}}
\usage{
as.curvacolina(x)
}
\arguments{
\item{x}{\code{data.frame}-like a ser convertido}
}
\description{
Forca um objeto tipo \code{data.frame} para classe \code{curvacolina}
}
\seealso{
Other curvacolina: 
\code{\link{geragrade}()},
\code{\link{learqcolina}()},
\code{\link{plot.curvacolina}()}
}
\concept{curvacolina}
