% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{geragrade}
\alias{geragrade}
\title{Gera Grade \code{nhl} x \code{npot} No Dominio Da Curva Colina}
\usage{
geragrade(colina, nhl, npot)
}
\arguments{
\item{colina}{objeto da classe \code{curvacolina} (retornado pelas funcoes de leitura)}

\item{nhl, npot}{numero de divisoes no eixo de queda liquida e potencia, respectivamente}
}
\value{
data.frame contendo coordenadas dos pontos na grade
}
\description{
Funcao auxiliar para gerar grades de pontos nos quais interpolar a curva colina
}
\examples{

grade <- geragrade(colinadummy, 20, 20)

\dontrun{
plot(colinadummy, "2d") + ggplot2::geom_point(data = grade, aes(hl, pot), col = 2)
}

}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{learqcolina}()},
\code{\link{plot.curvacolina}()}
}
\concept{curvacolina}
