% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getcolina}
\alias{getcolina}
\title{Generica Para Extrair Colina Original do Ajuste}
\usage{
getcolina(object)
}
\description{
Extrai o dado ajustado corretamente dependendo do tipo de modelo. Funcao interna
}
