% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-curvacolina.r
\name{learqcolina}
\alias{learqcolina}
\alias{learqprocit}
\title{Le Planilha de Curva Colina}
\usage{
learqcolina(arq, aba = 1)

learqprocit(arq)
}
\arguments{
\item{arq}{caminho da planilha}

\item{aba}{aba a ser lida. Por padrao igual a \code{1}, geralmente nao deve ser informado pelo
usuario}
}
\value{
objetos \code{curvacolina}: lista de um elemento \code{data.table} contendo as colunas

\describe{
\item{\code{hl}}{queda liquida}
\item{\code{pot}}{potencia gerada}
\item{\code{rend}}{rendimento correspondente}
}

\code{leaqrcolina} retorna apenas um objeto, enquanto \code{learqprocit} retornara uma lista de
tantos elementos quanto abas de curva colina existem na planilha
}
\description{
Funcoes para leitura de arquivos xlsx e extracao da curva colina em formato padronizado
}
\details{
\code{learqcolina} deve ser utilizada para leitura de arquivos especificos de curva colina, nos
quais existe apenas uma aba contendo a informacao das curvas. \code{learqprocit} e especifica
para uso com planilhas de processo iterativo, nas quais pode haver mais de uma curva colina.
}
\examples{

# arquivo de colina simples
arq <- system.file(package = "curvacolina", "extdata/colina.xlsx")
colina <- learqcolina(arq)

\dontrun{
plot(colina)
}

}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{geragrade}()},
\code{\link{plot.curvacolina}()}
}
\concept{curvacolina}
