% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.curvacolina}
\alias{plot.curvacolina}
\title{Visualizacao De \code{curvacolina}}
\usage{
\method{plot}{curvacolina}(x, tipo = c("3d", "2d"), print = TRUE, ...)
}
\arguments{
\item{x}{objeto \code{curvacolina} retornado por uma das funcoes de leitura}

\item{tipo}{um de \code{c("3d", "2d")} indicando o tipo de grafico desejado}

\item{print}{booleano indicando se o plot deve ser exibido. Caso \code{print = FALSE} o objeto
sera retornado silenciosamente}

\item{...}{existe somente para consistencia de metodos. Nao possui utilidade}
}
\value{
um objeto \code{plotly} contendo o plot. Em ambos os casos o grafico so sera exibido ao
usuario caso \code{print = TRUE} (o padrao).
}
\description{
Plots bi ou tridimensionais de curvas colinas
}
\examples{

arq <- system.file("extdata/colina.xlsx", package = "curvacolina")
colina <- learqcolina(arq)

\dontrun{
# plot 3d
plot(colina, "3d")
plot(colina, "2d")

# execucao silenciosa e posterior exibicao
p <- plot(colina, "2d", print = FALSE)
print(p)
}

}
\seealso{
Other curvacolina: 
\code{\link{as.curvacolina}()},
\code{\link{geragrade}()},
\code{\link{learqcolina}()},
\code{\link{summary.curvacolina}()},
\code{\link{write.curvacolina}()}
}
\concept{curvacolina}
