% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.interpolador}
\alias{plot.interpolador}
\title{Visualizacao De Interpolacoes}
\usage{
\method{plot}{interpolador}(x, tipo = c("3d", "2d"), add_colina = TRUE, print = TRUE, ...)
}
\arguments{
\item{x}{objeto \code{interpolador} retornado pela funcao homonima}

\item{tipo}{um de \code{c("3d", "2d")} indicando o tipo de grafico desejado}

\item{add_colina}{booleano indicando se os pontos da colina original tambem devem ser plotados}

\item{print}{booleano indicando se o plot deve ser exibido. Caso \code{print = FALSE} o objeto
sera retornado silenciosamente}

\item{...}{parametros passados para \code{\link{coordgrade}} para amostragem da superficie. Se
deixado vazio e gerada uma grade 200 x 200. Ver Detalhes}
}
\value{
Se \code{tipo = "3d"} um objeto \code{plotly} contendo o plot 3d; se \code{tipo = "2d"}
um objeto \code{ggplot} contendo o plot 2d. Em ambos os casos o grafico so sera exibido ao
usuario caso \code{print = TRUE} (o padrao).
}
\description{
Funcao para visualizacao dos modelos ajustados atraves de \code{interpolador}
}
\details{
Para plotar uma superficie e necessario amostrar pontos numa grade regular. A definicao desta
grade e realizada atraves do argumento \code{...}, que sera utilizado numa chamada de
\code{\link{coordgrade}} e esta grade amostrada do interpolador \code{x}. E possivel deixa-lo nao
especificado; neste caso sera usado o padrao \code{list(dhl = 200, dpot = 200)}, que deve ser
suficiente para um plot suave na maioria das circunstancias.
}
\seealso{
Other interpolador: 
\code{\link{interpolador}()},
\code{\link{predict.interpolador}()}
}
\concept{interpolador}
