% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacao.r
\name{predict.interpolador}
\alias{predict.interpolador}
\title{Previsao Com Modelos Interpoladores}
\usage{
\method{predict}{interpolador}(object, pontos, as.gradecolina, ...)
}
\arguments{
\item{object}{objeto da classe \code{interpolador} retornado pela funcao homonima}

\item{pontos}{data.frame ou matriz contendo pontos nos quais amostrar o rendimento}

\item{as.gradecolina}{booleano indicando o nivel de detalhe na saida da funcao. Ver Detalhes}

\item{...}{demais parametros que possam ser passados aos metodos de \code{predict} especificos}
}
\value{
vetor de rendimentos nas coordenadas especificadas em \code{pontos}
}
\description{
Metodo para interpolar curva colina a partir de um objeto gerado por \code{interpolador}
}
\details{
O argumento \code{as.gradecolina} permite alterar a saida da interpolacao. Se
\code{as.gradecolina = FALSE}, o padrao, apenas o vetor de rendimentos interpolados sera
retornado. Caso \code{as.gradecolina = TRUE}, a saida sera um objeto \code{\link{gradecolina}}.
}
\seealso{
Other interpolador: 
\code{\link{interpolador}()},
\code{\link{plot.interpolador}()}
}
\concept{interpolador}
