% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolacao.r
\name{predict.interpolador}
\alias{predict.interpolador}
\title{Previsao Com Modelos Interpoladores}
\usage{
\method{predict}{interpolador}(object, pontos, ...)
}
\arguments{
\item{object}{objeto da classe \code{interpolador} retornado pela funcao homonima}

\item{pontos}{data.frame ou matriz contendo pontos nos quais amostrar o rendimento}

\item{...}{demais parametros que possam ser passados aos metodos de \code{predict} especificos}

\item{plot}{booleano indicando se deve ser realizado um plot do resultado}
}
\value{
vetor de rendimentos nas coordenadas especificadas em \code{pontos}
}
\description{
Metodo para interpolar curva colina a partir de um objeto gerado por \code{interpolador}
}
\seealso{
Other interpolador: 
\code{\link{interpolador}()},
\code{\link{plot.interpolador}()}
}
\concept{interpolador}
