% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-tensorprod.r
\name{predict.tensorprod}
\alias{predict.tensorprod}
\title{Amostragem De Pontos Na Suavizacao}
\usage{
\method{predict}{tensorprod}(object, pontos, as.gradecolina = FALSE, ...)
}
\arguments{
\item{object}{objeto da classe \code{tensorprod} retornado pela funcao homonima}

\item{pontos}{data.frame ou matriz contendo coordenadas \code{(hl, pot)} dos pontos onde
interpolar}

\item{as.gradecolina}{booleano -- se \code{FALSE} (padrao) retorna apenas o vetor de rendimentos
interpolados nas coordenadas \code{pontos}; se \code{TRUE} um objeto \code{gradecolina}. Veja
\code{\link{gradecolina}}}

\item{...}{existe somente para consistencia de metodos. Nao possui utilidade}
}
\value{
vetor de rendimentos interpolados
}
\description{
Amostra as coordenadas especificadas via \code{pontos} na superficie suavizada
}
