% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-trangulacao.r
\name{predict.triangulacao}
\alias{predict.triangulacao}
\title{Amostragem De Pontos Na Triangulacao}
\usage{
\method{predict}{triangulacao}(object, pontos, ...)
}
\arguments{
\item{object}{objeto da classe \code{triangulacao} retornado pela funcao homonima}

\item{pontos}{data.frame ou matriz contendo coordenadas \code{(hl, pot)} dos pontos onde
interpolar}
}
\value{
vetor de rendimentos interpolados
}
\description{
Realiza interpolacao baricentrica de \code{pontos} nos triangulos da tesselacao
}
