% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-triangulacao.r
\name{triangulacao}
\alias{triangulacao}
\title{Triangulacao De \code{curvacolina}}
\usage{
triangulacao(colina, tessfunc = tessdelaunay)
}
\arguments{
\item{colina}{objeto \code{curvacolina} retornado pelas funcoes de leitura}
}
\value{
objeto da classe \code{triangulacao} contendo a tesselacao da curva colina
}
\description{
Funcao interna executada quando \code{metodo = "triangulacao"} em \code{interpolador}
}
\details{
Esta funcao nao deve ser chamada pelo usuario diretamente na maioria dos casos
}
